requirements.txt
# MiniQuantDeskV2 - PRODUCTION (Recommended)

pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

structlog>=23.2
python-json-logger>=2.0.7
rich>=13.7
colorama>=0.4.6

requests>=2.31
httpx>=0.27
aiohttp>=3.9
websocket-client>=1.8
websockets>=12.0

alpaca-py>=0.42
polygon-api-client>=1.15
yfinance>=0.2.40
finnhub-python>=2.4.20
alpha-vantage>=3.0.0

numpy>=2.1,<3
pandas>=2.2,<3
scipy>=1.12,<2
pandas-ta>=0.4.67b0
ta>=0.11.0

schedule>=1.2.0
psutil>=5.9
python-dateutil>=2.8.2
tzdata>=2024.1

discord-webhook>=1.3.0

openpyxl>=3.1.2
pyarrow>=12.0.0
